backend/rea-telemetry/telemetry/jobs/1_fetch_telemetry_data_job/requirements.txt
functions-framework==3.*

google-cloud-firestore==2.20.1
google-cloud-logging==3.12.0
google-cloud-pubsub==2.29.0
google-cloud-secret-manager==2.23.2
orjson==3.10.16
pydantic==2.11.3
python-dateutil==2.9.0
requests==2.32.3
tenacity==9.1.2
